requirements.txt
obs-websocket-py
Flask